refactor: Refactor model
